Add staff check to invite detection
And some other stuff

+++ cogs/automod.py
@@ -1,5 +1,7 @@
 import re
 
+import discord
+
 INVITE_REGEX = "(https?:\/\/)?(www\.)?(discord\.(gg|io|me|li)|discordapp\.com\/invite)\/.+"
 
 
@@ -11,11 +13,16 @@ class Automod:
         self.inviteregex = re.compile(INVITE_REGEX)
 
     async def on_message(self, msg):
+        if msg.guild is None:
+            return
+
         # Check if message content contains an invite url
         invite = self.inviteregex.search(msg.content)
         if invite:
-            await msg.delete()
-            await msg.channel.send("No")
+            # Ignore staff though
+            if discord.utils.get(msg.author.roles, name="Staff") is None:
+                await msg.delete()
+                await msg.channel.send(f"{msg.author.mention} you sent an invite link, I deleted it for you.")
 
 
 def setup(bot):
